pyproject.toml
[tool.poetry]
name = "crane-users"
version = "0.1.0"
description = ""
authors = ["poleshkomi <poleshkomi@modeltech.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.2"
python-dotenv = "^1.0.1"
alembic = "^1.13.3"
pydantic-settings = "^2.6.0"
asyncpg = "^0.30.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
cerberus = "^1.3.5"
bcrypt = "^4.2.0"
typer = "^0.13.0"
uvicorn = "^0.32.0"
pyjwt = "^2.10.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.0"
mypy = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
